app/requirements.txt
fastapi
uvicorn
pytest
httpx
python-dotenv

# ML Dependencies
transformers>=4.36.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
torch>=2.1.0
sentencepiece
protobuf
datasets
peft>=0.7.0

# RAG Dependencies
tavily-python

# Content Moderation
better-profanity

# Semantic Routing & Metrics
sentence-transformers>=2.2.0
prometheus-client>=0.19.0
sse-starlette>=1.6.0
ollama>=0.1.0
scikit-learn>=1.3.0

